feature: Replace jq expressions with CEL for team and group ref extraction
To implement dynamic field selection with jq queries in
GenericPlatformTeamsFetcher we've used jackson-jq library.
It appears not to be actively maintained and it's the only dependency
which stops us from upgrading to java25, spring 7, spring boot 4 and
jackson 3.
Given that, it make sense to migrate to something more actively
maintained and with less external dependencies - google-cel