feature: Replace jq expressions with CEL for team and group ref extraction (#129)
* feature: Replace jq expressions with CEL for team and group ref extraction

To implement dynamic field selection with jq queries in
GenericPlatformTeamsFetcher we've used jackson-jq library.
It appears not to be actively maintained and it's the only dependency
which stops us from upgrading to java25, spring 7, spring boot 4 and
jackson 3.
Given that, it make sense to migrate to something more actively
maintained and with less external dependencies - google-cel

* feature: Do not fail the bot in case CEL evaluation is fails

It can be an unexpected change in the environment, we don't want to make
the bot stop working because of this

* feature: Replace jq expressions with CEL for team and group ref extraction

To implement dynamic field selection with jq queries in
GenericPlatformTeamsFetcher we've used jackson-jq library.
It appears not to be actively maintained and it's the only dependency
which stops us from upgrading to java25, spring 7, spring boot 4 and
jackson 3.
Given that, it make sense to migrate to something more actively
maintained and with less external dependencies - google-cel

* feature: Do not fail the bot in case CEL evaluation is fails

It can be an unexpected change in the environment, we don't want to make
the bot stop working because of this

* feature: Fix tests